Parser can now read/write data from/to labels.
A node with a root set of state 1 for the first character,
state 2 for second, etc, can be specified as a string:
A[[1|2|3][][]]

A node with single character and root set of A,B
and VU of AB, and VL set of C can be specified as
A[[A,B][A,B][C]]

The two can be combined. Note that the parser doesn't
ensure that the sets are actually possible VV/VU/VL. It is
unaffiliated with the algorithms which act upon that data.